school/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
Markdown==3.3.4
PyJWT==2.3.0
pytz==2021.1
sqlparse==0.4.2
urllib3==1.26.7

